api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pypdf==5.0.0
pikepdf==9.3.0
python-multipart==0.0.9
reportlab==4.2.5
pillow==10.4.0
